+++ src/stapi_fastapi/constants.py
@@ -1,3 +1,2 @@
-STAPI_VERSION = "0.0.0.pre"
 TYPE_JSON = "application/json"
 TYPE_GEOJSON = "application/geo+json"