+++ tests/test_opportunity.py
@@ -1,5 +1,3 @@
-from datetime import UTC, datetime, timedelta
-
 import pytest
 from fastapi.testclient import TestClient
 
@@ -8,40 +6,15 @@ from stapi_fastapi.models.opportunity import (
 )
 
 from .shared import create_mock_opportunity, pagination_tester
-from .test_datetime_interval import rfc3339_strftime
 
 
-@pytest.mark.parametrize("product_id", ["test-spotlight"])
 def test_search_opportunities_response(
-    product_id: str,
-    stapi_client: TestClient,
-    assert_link,
+    stapi_client: TestClient, assert_link, opportunity_search
 ) -> None:
-    now = datetime.now(UTC)
-    end = now + timedelta(days=5)
-    format = "%Y-%m-%dT%H:%M:%S.%f%z"
-    start_string = rfc3339_strftime(now, format)
-    end_string = rfc3339_strftime(end, format)
-
-    request_payload = {
-        "geometry": {
-            "type": "Point",
-            "coordinates": [0, 0],
-        },
-        "datetime": f"{start_string}/{end_string}",
-        "filter": {
-            "op": "and",
-            "args": [
-                {"op": ">", "args": [{"property": "off_nadir"}, 0]},
-                {"op": "<", "args": [{"property": "off_nadir"}, 45]},
-            ],
-        },
-        "limit": 10,
-    }
-
+    product_id = "test-spotlight"
     url = f"/products/{product_id}/opportunities"
 
-    response = stapi_client.post(url, json=request_payload)
+    response = stapi_client.post(url, json=opportunity_search)
 
     assert response.status_code == 200, f"Failed for product: {product_id}"
     body = response.json()
@@ -61,6 +34,7 @@ def test_search_opportunities_response(
 def test_search_opportunities_pagination(
     limit: int,
     stapi_client: TestClient,
+    opportunity_search,
 ) -> None:
     mock_pagination_opportunities = [create_mock_opportunity() for __ in range(3)]
     stapi_client.app_state["_opportunities"] = mock_pagination_opportunities
@@ -71,29 +45,6 @@ def test_search_opportunities_pagination(
             x.model_dump(mode="json") for x in mock_pagination_opportunities
         ]
 
-    now = datetime.now(UTC)
-    start = now
-    end = start + timedelta(days=5)
-    format = "%Y-%m-%dT%H:%M:%S.%f%z"
-    start_string = rfc3339_strftime(start, format)
-    end_string = rfc3339_strftime(end, format)
-
-    request_payload = {
-        "geometry": {
-            "type": "Point",
-            "coordinates": [0, 0],
-        },
-        "datetime": f"{start_string}/{end_string}",
-        "filter": {
-            "op": "and",
-            "args": [
-                {"op": ">", "args": [{"property": "off_nadir"}, 0]},
-                {"op": "<", "args": [{"property": "off_nadir"}, 45]},
-            ],
-        },
-        "limit": limit,
-    }
-
     pagination_tester(
         stapi_client=stapi_client,
         url=f"/products/{product_id}/opportunities",
@@ -101,5 +52,5 @@ def test_search_opportunities_pagination(
         limit=limit,
         target="features",
         expected_returns=expected_returns,
-        body=request_payload,
+        body=opportunity_search,
     )